wait_for_stop
